requirements.txt
GitPython==0.3.2.RC1
lxml==3.1.0